rename scripts in services
