+++ package.json
@@ -63,5 +63,5 @@
     "prep": [
       "build"
     ]
-  },
+  }
 }
