+++ src/app/wizard/services/animation.factory.js
