+++ src/app/wizard/services/platform.js
